requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
APScheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
aiofiles==23.2.1